cli: add --json and --experimental-json flags, split JSONL event processor
Split event_processor_with_json_output into two implementations: existing one for --json, and new event_processor_with_experimental_json_output providing improved, testable JSONL output for --experimental-json. Add mutually exclusive --json and --experimental-json CLI flags and update wiring accordingly. Deprecation message is shown when --json is used. Tests updated to use experimental implementation.
